Wallrun V2
Changes :
- Most of the logic has been moved to SO_CharacterMovementComponent.
- Fixed various bugs.

Bugs :
- Tend to "snap" to small walls, maybe extend the detection system.

Comment :
Will not be sturdy enough for a full project but work well for a prototype.
Should implement a new movement state in the movement component, will take too much time.
For now the falling state is "overrided" when player is wall runing.
