+++ streumon_fps/Content/Mannequin/Animations/ThirdPerson_AnimBP.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9643a88fd2a2a17f2c6188e12fe5c541c51512a06b70de7c286cc852352004fe
-size 329982
+oid sha256:86ceb65a63282f1f69a3f74e556c9c4861078e7d9c7e8664c53ef40444f324b5
+size 333086
