+++ streumon_fps/Content/ThirdPersonCPP/Blueprints/ThirdPersonCharacter.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3b8fe023ded5729f2dd4d7c2791426702ef66206954ba3206cd4e3ce9545205f
-size 212443
+oid sha256:e4483ad9b4d992ce255f8b6d4cc53bbf12d5fcd70149866495fc113cbad24344
+size 213067
