+++ streumon_fps/Content/ThirdPersonCPP/Maps/ThirdPersonExampleMap.umap
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:13ccfdad6079ec3bf60940c4cf4265a9c1c4188c59e7eea631a84855dd3f76e4
-size 8527078
+oid sha256:d8d09979c099711e6a1018ded7d37aa9aa968c0a6560400501ef5ba0fe8cf92e
+size 8531098
