+++ streumon_fps/Content/ThirdPersonCPP/Maps/ThirdPersonExampleMap_BuiltData.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b21493e2713f8ac9a54ca68164b25f41c76a76cafc868436b309e35f6fce6342
+oid sha256:818e58a25dc51362124c99eef4c8108ff3bfae0ae039ddbe19778f4af6337c69
 size 3393394
