+++ streumon_fps/clean.sh
@@ -1,0 +1,6 @@
+# Light script for code cleaning
+rm SquadLeader.sln
+rm -r .vs/
+rm -r Intermediate/
+rm -r Binaries/
+rm -r DerivedDataCache/
